build scripts refactor
- renamed helper -> updateFileAsync
- checks around directory removals
- long paths param added to clone calls
- removed explicit build flavour (not being used) flavour in name is derived from working directory
- removed obsolete script modification steps
- do not zip output if build fails
- append UE and SDK versions automatically to output
- readFileValueAsync helper script
- cleanup scripts
