+++ VoiceSDK_1click_install_scripts/install_VoiceSDK.bat
@@ -3,12 +3,12 @@
 :: This source code is licensed under the license found in the
 :: LICENSE file in the root directory of this source tree.
 
-mkdir Plugins
+if not exist Plugins\ mkdir Plugins
 cd Plugins
-git clone https://github.com/wit-ai/voicesdk-unreal
-git clone https://github.com/wit-ai/wit-unreal
+
+if not exist voicesdk-unreal\ git clone -c core.longpaths=true https://github.com/wit-ai/voicesdk-unreal
+if not exist wit-unreal\ git clone -c core.longpaths=true https://github.com/wit-ai/wit-unreal
 
 cd ..
-
 ECHO [Voice]>>.\\Config\\DefaultEngine.ini
 ECHO bEnabled=true>>.\\Config\\DefaultEngine.ini
