+++ scripts/all-in-one/TP_4/cleanup.bat
@@ -5,4 +5,4 @@
 
 echo off
 
-../utils/all_in_one_packaging.bat 4.27
+../utils/cleanup.bat
