+++ scripts/all-in-one/TP_5_1/build.bat
@@ -5,4 +5,4 @@
 
 echo off
 
-../utils/all_in_one_packaging.bat 5.1 TP
+../utils/all_in_one_packaging.bat 5.1
