+++ scripts/all-in-one/UE5_1/cleanup.bat
@@ -5,4 +5,4 @@
 
 echo off
 
-../utils/all_in_one_packaging.bat 4.27
+../utils/cleanup.bat
