+++ scripts/all-in-one/github_UE4/all_in_one_packaging.bat
@@ -3,30 +3,33 @@
 :: This source code is licensed under the license found in the
 :: LICENSE file in the root directory of this source tree.
 
-rmdir /s /q output
+if exist output\ rmdir /s /q output
 mkdir output
-rmdir /s /q workspace
+
+if exist workspace\ rmdir /s /q workspace
 mkdir workspace
 
 cd workspace
 
-git clone https://github.com/wit-ai/voicesdk-unreal
-git clone https://github.com/wit-ai/wit-unreal
-
-node ../../update_WITH_VOICESDK_USERAGENT_flag.js
-node ../../utils/update_plugin_flavour.js aio
+git clone -c core.longpaths=true https://github.com/wit-ai/voicesdk-unreal
+git clone -c core.longpaths=true https://github.com/wit-ai/wit-unreal
 
 cd voicesdk-unreal
 rmdir /s /q .git
 rmdir /s /q scripts
 
-cd ..
-cd wit-unreal
+cd ..\wit-unreal
 rmdir /s /q .git
 rmdir /s /q scripts
 
-cd ..
+set currPath=%CD%
+cd ../../../utils
+for /f %%i in ('node read_Plugin_version.js') do set sdkVersion=%%i
+cd %currPath%
 
-powershell Compress-Archive -Path wit-unreal, voicesdk-unreal -DestinationPath ../output/voicesdk-all-in-one-ue4.zip
 cd ..
-echo "The UE4 VoiceSDK all-in-one plugin is ready in output/voicesdk-all-in-one-ue4.zip"
+set outputName=voicesdk-all-in-one-UE4-v%sdkVersion%.zip
+powershell Compress-Archive -Path wit-unreal, voicesdk-unreal -DestinationPath ../output/%outputName%
+
+cd ..
+echo "The UE4 VoiceSDK all-in-one plugin is ready in output/%outputName%"
