--- scripts/all-in-one/update_WITH_VOICESDK_USERAGENT_flag.js
@@ -1,16 +1,0 @@
-/*
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-var fs = require('fs');
-const updateFileAsync = require("../helper");
-
-const FILE_PATH = 'wit-unreal\\Source\\Wit\\Wit.Build.cs';
-const KEY_STRING_START = 'PrivateDefinitions.Add("WITH_VOICESDK_USERAGENT=';
-const KEY_STRING_END = '"';
-const replacement = '1'
-
-updateFileAsync(FILE_PATH, KEY_STRING_START, replacement, KEY_STRING_END);
