+++ scripts/updateFileAsync.js
